* master: (554 commits)
  [WebProfilerBundle] fixed typo
  [WebProfileBundle] remove dependency on the DIC
  [WebProfilerBundle] removed dependency on the templating component
  [HttpKernel] unified the way the traceable event dispatcher injects information into the profiler (closes #5733)
  [WebProfilerBundle] avoid the usage of the global app variable
  [WebProfilerBundle] changed all includes to use the new Twig's notation
  [WebProfilerBundle] moved all static assets directly into the templates
  [Console] [ProgressHelper] better percentage precision
  [Form] Removed unused method ChoiceView::isSelected()
  Fix incorrect inheritdoc blocks
  [DomCrawler] fixed a bad merge
  [Console] Fixes in ProgressHelper
  [Security] Added Pbkdf2PasswordEncoder
  FIX [2.1][ClassLoader]UniversalClassLoader not working with AnnotationRegistry::registerLoader
  fixed typo
  [Config] added some phpdocs
  added entry to changelog
  fixed CS
  adds two convenience methods for optional configuration sections
  added doc comments
  ...

Conflicts:
	src/Symfony/Component/HttpFoundation/Tests/StreamedResponseTest.php

+++ src/Symfony/Component/HttpFoundation/Tests/Session/Storage/MockFileSessionStorageTest.php
@@ -28,7 +28,7 @@ class MockFileSessionStorageTest extends \PHPUnit_Framework_TestCase
     private $sessionDir;
 
     /**
-     * @var MockFileSessionStorage
+     * @var FileMockSessionStorage
      */
     protected $storage;
 
@@ -40,14 +40,12 @@ class MockFileSessionStorageTest extends \PHPUnit_Framework_TestCase
 
     protected function tearDown()
     {
-        foreach (glob($this->sessionDir.'/mocksess_*') as $file) {
-            unlink($file);
-        }
-
-        rmdir($this->sessionDir);
-
         $this->sessionDir = null;
         $this->storage = null;
+        array_map('unlink', glob($this->sessionDir.'/*.session'));
+        if (is_dir($this->sessionDir)) {
+            rmdir($this->sessionDir);
+        }
     }
 
     public function testStart()
